Added more unit tests for Fringe2WrappedPhase. Also found a bug where we were passing the pitch as a float

+++ PortalCapture/src/Shaders/Fringe2WrappedPhase.frag
@@ -2,8 +2,8 @@
 
 uniform sampler2D fringeImage1;
 uniform sampler2D fringeImage2;
-uniform float pitch1;
-uniform float pitch2;
+uniform int pitch1;
+uniform int pitch2;
 
 in vec2 fragTexCoord;
